Fix buffer sizes used for map names (64 -> MAX_PATH).
As more games are now supporting maps nested in subfolders or in folders outside of the maps folder,
we need to account for the full path that the game uses to refer to the map for compatibility. Many other
places for fixed for this already after CS:GO added Steam Workshop support for maps.