Adds a flag to RequestFrame to specify a Think hook
Added AddThinkAction to the sm interface and a flag to RequestFrame on
pawn to specify a Think hook instead of a GameFrame one. This allows
plguins to schedule actions to be executed on the next think call, even
if the server is hibernating.

Status: the branch builds no problem but it generates a runtime error
while trying to load the sdktools and sdkhooks extensions ("Unable to
load extension "sdkhooks.ext": undefined symbol: Warning" and "Unable to
load extension "sdktools.ext": undefined symbol: Warning") any help
would be appreciated.