Fix crash when database transaction fails (bug 6531)
If one of the queries fails in a transaction, the DBI system would only
allocate an array of the size of the amount of all the successful
queries before the failed one.
It writes data for all the queries though effectively writing past the
array bounds leading to heap corruption.

Create the right sized array!