Use pseudo addresses on 64-bit platforms in places where natives use memory addresses.
This is a stop-gap solution until SourcePawn is able to handle 64-bit integers natively.
Pseudo addresses are made up of a 6-bit table index and a 26-bit offset.
The table consists of base addresses to memory allocations.
Using a scheme like this allows a limited amount of pointer arithmetic to be performed.