Updated formatting for consistency.
This PR is perhaps overkill, however this is something I've wanted to get to for a few months. 95% of the changes are moving from tabs to spaces for aligning text (not initial indents, which have remained as tabs)

Other misc changes:
-Added missing @deprecated comments
-Modified braces on stock functions for consistency
-Removed empty spaces at the end of lines,
-Removed commas following last enum member.
-A bunch of other minor junk that I can't remember

This PR is probably pretty obnoxious.

Most tab to space conversions were done through regexes:
```
((?<!^)\t)|(?<= )\t

param ([\w\.]{1})\t+
param \1

@return[ \t]+
@return

@error[ \t]+
@error

\*[ \t]{3,}
*

[ \t]+$
```

Everything else was done manually with the use of some beer and music.