Fix regex inc docs
https://steamuserimages-a.akamaihd.net/ugc/786360440444926021/3977E37BB5937C22C7C82A96DC503ED449F865B1/

Some methodmap documentation was a direct copypasta from the non-methodmap natives, mentioning a handle parameter which doesn't exist (since the "this" variable is the handle itself). A couple non-methodmap natives had the @param documentation in the wrong order, with the handle coming second.

I think I fixed everything in this file. Please review and let me know of any changes that should be made. Thanks.