Allow to escape string even if buffer is not 2*strlen(string)+1.
The problem with the actual code that if the buffer is not size x 2 +1 the function returns false even if it was possible to escape string.

I suggest this implementation to allow this case. First call the basic function mysql_real_escape_string() then test if needed is in range of maxlength. The return of mysql_real_escape_string() return -1 if failure and allow the native to return false.