mapchooser: Set minimum number maps to include in vote
This commit allows the user to have a consistent number minimum number of maps in the vote, regardless how many previous maps are excluded. It also improves the handling of the situation, when the number is previous maps to exclude is set higher than the number of available maps. Instead of excluding no previous maps, it will exclude until there's a configurable number of minimum options left.