Add native IsLateLoadTime
Adding this to reduce the amount of code required in a plugin to perform an action if the plugin was loaded late as this can be a general problem for handling a plugin when it is reloaded.
This saves having to store the late load state your self from the `AskPluginLoad2` forward.

Currently:
g_bLateLoaded = false;

public APLRes AskPluginLoad2(Handle plugin, bool late, char[] error, int err_max)
{
    g_bLateLoaded = late;
}

public void OnPluginStart()
{
    if (g_bLateLoaded)
    {
        // Do stuff on late load
    }
}

After this addition:
public void OnPluginStart()
{
    if (IsLateLoadTime())
    {
        // Do stuff on late load
    }
}