Fix vprof crashing in some games (#1541)
Some engines are very sensitive to exactly when in a frame vprof is
enabled, the vprof commands use a special command registration method
to defer their execution to the start of the next frame. Instead of
starting/stopping vprof directly ourselves, use the built-in commands
to ensure that the timing is correct and the server does not crash.

Fixes #1162

(cherry picked from commit f8f5a18d675be3284170bf2900ec2194b0e2d33c)