Return array type info with FindSendPropInfo
FindSendPropInfo currently isn't very useful with array props, it'll
return the correct offset but the type will be unknown and the bit count
zero. Instead, populate type and num_bits with the underlying array type
info same as we'd handle it when reading/writing the prop. Closes #1258.

This also adds the preliminary for Source's legacy / virtual array props
for issue #1386. They can now be looked up correctly with
FindSendPropInfo and it returns the correct type / size / offset.

Co-Authored-By: nosoop <nosoop@users.noreply.github.com>