DHooks: Fix changing of byref vector parameters
We always created a new vector object instead of changing the passed in vector directly instead. This works for the function being called using our changed values - but the caller doesn't see the change values if it's passing a vector by reference.

Only create a new vector if there isn't one being passed in and set the values directly in the passed in vector otherwise.