DHooks: Remove function entry on a new matching section
When two separate gamedata files are loaded in and one wants to
declare a function name as a detour and another wants to declare it
as a virtual hook, any `DHookSetup.SetFromConf` operations on that
name after the second gamedata file is parsed will fail as the
`SignatureWrapper` container contains references to both
offsets / signatures.

This PR changes the behavior so new matching functions fully
override previous ones.  This does mean `DHookSetup.SetFromConf`
needs to be called before other gamedata files are parsed to ensure
the hook is initialized as expected.